Fix bug with sprinting and tools
When a tool was dropped while sprinting then player would move
with default speed while holding shift, and slower than normal
without.

Changed the harcoded values for speed with and without tool to
a speed modifier, which is subtracted or added to the current
speed. This works well with sprinting.